backend/requirements.txt
flask
paho-mqtt
sqlalchemy